requirements.txt
apache-airflow==3.1.8
apache-airflow-providers-google==20.0.0